micro-services/auth/requirements.txt
Flask==3.1.3
flask-cors==6.0.1
requests==2.31.0
firebase-admin==6.5.0
python-dotenv==1.1.1
pytest==8.3.5
